Adjust deploy so everything can build and server runs.
Add some things to .gitignore.

Update the api so that it builds modules so that the client is willing to import it.

Update the server so that it is run through webpack which is more forgiving about importing modules (to work around nodejs native module support failing when the api is built in a way that the client).

Fix some imports in the api to be of .js because TypeScript does not change strings on imports because TypeScript is confusing.

Update the server so that it listens to the PORT environment variable when deciding on which port to listen to.